requirements.txt
fastapi>=0.42.0
uvicorn>=0.9.1
redis>=3.3.11
python-dotenv>=0.10.3
pytest>=5.1.2